+++ src/os2datascanner/engine2/tests/test_messages.py
@@ -104,3 +104,26 @@ class TestMessage:
         # Assert
         assert isinstance(mo, messages.ContentMissingMessage)
         assert mo.handle.relative_path == "path/to/document.txt"
+
+    def test_problem_irrelevant_compat(self):
+        """Old-fashioned ContentIrrelevantMessages (based on ProblemMessage)
+        can be parsed."""
+        # Arrange
+        jf = {
+            "scan_tag": messages.ScanTagFragment.make_dummy().to_json_object(),
+            "handle": {
+                "type": "file",
+                "source": {
+                    "type": "file",
+                    "path": "/home/af",
+                },
+                "path": "path/to/second-document.txt",
+            },
+            "irrelevant": True,
+            "message": "Not part of the scan no more"
+        }
+        # Act
+        mo = messages.ProblemMessage.from_json_object(jf)
+        # Assert
+        assert isinstance(mo, messages.ContentIrrelevantMessage)
+        assert mo.handle.relative_path == "path/to/second-document.txt"